requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.2
plotly==5.22.0
streamlit==1.33.0
